Move examples to a new folder
