+++ src/examples/alerts/AlertExample1.tsx
@@ -1,0 +1,19 @@
+import VAlert from '@/components/VAlert';
+import { component } from 'vue-tsx-support';
+import SubTitle from '../SubTitle';
+
+const AlertExample1 = component({
+  render() {
+    return (
+      <div>
+        <SubTitle>Alert</SubTitle>
+        <VAlert color="success" value={true}>
+          Awesome!
+        </VAlert>
+        <VAlert value={true}>Oh !@#$!</VAlert>
+      </div>
+    );
+  },
+});
+
+export default AlertExample1;
